[MIG] base_partner_one2many_phone v14 -> v16
