+++ base_partner_one2many_phone/__init__.py
@@ -1,2 +1,2 @@
-from . import partner_phone
+from . import models
 from .post_install import migrate_to_partner_phone
