+++ base_partner_one2many_phone/__manifest__.py
@@ -1,12 +1,12 @@
-# Copyright 2014-2020 Abbaye du Barroux (http://www.barroux.org)
-# Copyright 2014-2020 Akretion (http://www.akretion.com>)
+# Copyright 2014-2023 Abbaye du Barroux (http://www.barroux.org)
+# Copyright 2014-2023 Akretion (http://www.akretion.com>)
 # @author: Frère Bernard <informatique@barroux.org>
 # @author: Alexis de Lattre <alexis.delattre@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
     'name': 'Base Partner One2many Phone',
-    'version': '14.0.1.0.0',
+    'version': '16.0.1.0.0',
     'category': 'Phone',
     'license': 'AGPL-3',
     'summary': 'One2many link between partners and phone numbers/emails',
@@ -19,13 +19,14 @@ With this module, one partner can have several phone numbers and several emails.
 It has been developped by brother Bernard from Barroux Abbey and Alexis de Lattre from Akretion.
     """,
     'author': 'Akretion',
-    'website': 'https://akretion.com/',
+    'website': 'https://github.com/akretion/odoo-usability',
     'depends': ['contacts', 'base_usability', 'phone_validation'],
     'excludes': ['sms'],  # because sms introduces big changes in partner form view
     'data': [
-        'partner_phone_view.xml',
+        'views/res_partner_phone.xml',
+        'views/res_partner.xml',
         'security/ir.model.access.csv',
         ],
-    'installable': False,
+    'installable': True,
     'post_init_hook': 'migrate_to_partner_phone',
 }
