+++ base_partner_one2many_phone/models/__init__.py
@@ -1,0 +1,2 @@
+from . import res_partner_phone
+from . import res_partner
